requirements.txt
absl-py==2.2.2
accelerate==1.6.0
annotated-types==0.7.0
asttokens==3.0.0
av==14.3.0
bitmath==1.3.3.1
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
decord==0.6.0
deprecation==2.1.0
dill==0.3.9
docker-pycreds==0.4.0
einops==0.8.1
enum-tools==0.12.0
etils==1.12.2
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
graphviz==0.20.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
hbutils==0.11.0
huggingface-hub==0.30.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jsonargparse==4.38.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lmdb==1.6.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
mujoco==3.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
openrl==0.2.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.2
pydantic_core==2.33.1
pygame==2.6.1
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.25.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
termcolor==3.0.1
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
treevalue==1.5.0
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.9
wcwidth==0.2.13
zipp==3.21.0
